replaced usize

+++ secp256k1-sys/src/lib.rs
@@ -456,17 +456,17 @@ extern "C" {
     pub fn secp256k1_rangeproof_rewind(
         cx: *const Context,
         blind_out: *mut c_uchar,
-        value_out:  *mut size_t,//longlong
+        value_out:  *mut u64,//longlong
         message_out:  *mut c_uchar,
         outlen: *mut size_t,
         nonce:  *mut c_uchar,
-        min_value:  *mut size_t, //longlong
-        max_value:  *mut size_t, // longlong
+        min_value:  *mut u64, //longlong
+        max_value:  *mut u64, // longlong
         commit: *mut c_uchar, // rustsecp256k1_v0_4_1_pedersen_commitment
         proof: *mut c_uchar,
-        plen: size_t, // without *
+        plen: u64, // without *
         extra_commit: *mut c_uchar,
-        extra_commit_len: size_t, //without * 
+        extra_commit_len: u64, //without * 
         //gen: *mut c_uchar // generator
     ) -> c_int;
